+++ sdk/data/aztables/client.go
@@ -8,11 +8,11 @@ import (
 	"encoding/json"
 	"errors"
 	"net/url"
-	"reflect"
 	"strings"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	generated "github.com/Azure/azure-sdk-for-go/sdk/data/aztables/internal"
 )
@@ -106,16 +106,19 @@ func parseURL(serviceURL string) (string, string, error) {
 
 // CreateTableResponse contains response fields for Client.Create and ServiceClient.CreateTable
 type CreateTableResponse struct {
-	// placeholder for future response fields
+	// The name of the table.
+	TableName *string `json:"TableName,omitempty"`
 }
 
 func createTableResponseFromGen(g *generated.TableClientCreateResponse) CreateTableResponse {
-	return CreateTableResponse{}
+	return CreateTableResponse{
+		TableName: g.TableName,
+	}
 }
 
-// Create creates the table with the tableName specified when NewClient was called. If the service returns a non-successful
+// CreateTable creates the table with the tableName specified when NewClient was called. If the service returns a non-successful
 // HTTP status code, the function returns an *azcore.ResponseError type. Specify nil for options if you want to use the default options.
-func (t *Client) Create(ctx context.Context, options *CreateTableOptions) (CreateTableResponse, error) {
+func (t *Client) CreateTable(ctx context.Context, options *CreateTableOptions) (CreateTableResponse, error) {
 	if options == nil {
 		options = &CreateTableOptions{}
 	}
@@ -141,11 +144,11 @@ type ListEntitiesOptions struct {
 	// Maximum number of records to return.
 	Top *int32
 
-	// The PartitionKey to start paging from
-	PartitionKey *string
+	// The NextPartitionKey to start paging from
+	NextPartitionKey *string
 
-	// The RowKey to start paging from
-	RowKey *string
+	// The NextRowKey to start paging from
+	NextRowKey *string
 }
 
 func (l *ListEntitiesOptions) toQueryOptions() *generated.QueryOptions {
@@ -155,91 +158,43 @@ func (l *ListEntitiesOptions) toQueryOptions() *generated.QueryOptions {
 
 	return &generated.QueryOptions{
 		Filter: l.Filter,
-		Format: generated.ODataMetadataFormatApplicationJSONODataMinimalmetadata.ToPtr(),
+		Format: to.Ptr(generated.ODataMetadataFormatApplicationJSONODataMinimalmetadata),
 		Select: l.Select,
 		Top:    l.Top,
 	}
 }
 
-// ListEntitiesPageResponse contains response fields for ListEntitiesPager.NextPage
-type ListEntitiesPageResponse struct {
-	// ContinuationNextPartitionKey contains the information returned from the x-ms-continuation-NextPartitionKey header response.
-	ContinuationNextPartitionKey *string
+// ListEntitiesResponse contains response fields for ListEntitiesPager.NextPage
+type ListEntitiesResponse struct {
+	// NextPartitionKey contains the information returned from the x-ms-continuation-NextPartitionKey header response.
+	NextPartitionKey *string
 
-	// ContinuationNextRowKey contains the information returned from the x-ms-continuation-NextRowKey header response.
-	ContinuationNextRowKey *string
-
-	// The metadata response of the table.
-	ODataMetadata *string
+	// NextRowKey contains the information returned from the x-ms-continuation-NextRowKey header response.
+	NextRowKey *string
 
 	// List of table entities.
 	Entities [][]byte
 }
 
 // transforms a generated query response into the ListEntitiesPaged
-func newListEntitiesPage(resp generated.TableClientQueryEntitiesResponse) (ListEntitiesPageResponse, error) {
+func newListEntitiesPage(resp generated.TableClientQueryEntitiesResponse) (ListEntitiesResponse, error) {
 	marshalledValue := make([][]byte, 0)
 	for _, e := range resp.TableEntityQueryResponse.Value {
 		m, err := json.Marshal(e)
 		if err != nil {
-			return ListEntitiesPageResponse{}, err
+			return ListEntitiesResponse{}, err
 		}
 		marshalledValue = append(marshalledValue, m)
 	}
 
-	return ListEntitiesPageResponse{
-		ContinuationNextPartitionKey: resp.XMSContinuationNextPartitionKey,
-		ContinuationNextRowKey:       resp.XMSContinuationNextRowKey,
-		ODataMetadata:                resp.TableEntityQueryResponse.ODataMetadata,
-		Entities:                     marshalledValue,
+	return ListEntitiesResponse{
+		NextPartitionKey: resp.XMSContinuationNextPartitionKey,
+		NextRowKey:       resp.XMSContinuationNextRowKey,
+		Entities:         marshalledValue,
 	}, nil
 }
 
-// ListEntitiesPager is a Pager for Table entity query results.
-//
-// NextPage should be called first, it fetches the next available page of results from the service.
-// If the fetched page contains results, the return value is true, else false.
-// Results fetched from the service can be evaluated by calling PageResponse on this Pager.
-// If the result is false, the value of Err() will indicate if an error occurred.
-//
-// PageResponse returns the results from the page most recently fetched from the service.
-type ListEntitiesPager struct {
-	tableName   string
-	client      *generated.TableClient
-	current     generated.TableClientQueryEntitiesResponse
-	listOptions *ListEntitiesOptions
-	nextPK      *string
-	nextRK      *string
-}
-
-// More returns true if there are more pages to fetch, false otherwise.
-func (p *ListEntitiesPager) More() bool {
-	if !reflect.ValueOf(p.current).IsZero() {
-		if p.current.XMSContinuationNextPartitionKey == nil || len(*p.current.XMSContinuationNextPartitionKey) > 0 || p.current.XMSContinuationNextRowKey == nil || len(*p.current.XMSContinuationNextRowKey) > 0 {
-			return false
-		}
-	}
-	return true
-}
-
-// NextPage fetches the next available page of results from the service.
-// If the fetched page contains results, the return value is true, else false.
-// Results fetched from the service can be evaluated by calling PageResponse on this Pager.
-func (p *ListEntitiesPager) NextPage(ctx context.Context) (ListEntitiesPageResponse, error) {
-	resp, err := p.client.QueryEntities(ctx, generated.Enum1Three0, p.tableName, &generated.TableClientQueryEntitiesOptions{
-		NextPartitionKey: p.nextPK,
-		NextRowKey:       p.nextRK,
-	}, p.listOptions.toQueryOptions())
-	if err != nil {
-		return ListEntitiesPageResponse{}, err
-	}
-	p.current = resp
-	p.nextPK = resp.XMSContinuationNextPartitionKey
-	p.nextRK = resp.XMSContinuationNextRowKey
-	return newListEntitiesPage(resp)
-}
-
-// List queries the entities using the specified ListEntitiesOptions.
+// ListEntities queries the entities using the specified ListEntitiesOptions.
 // ListEntitiesOptions can specify the following properties to affect the query results returned:
 //
 // Filter: An OData filter expression that limits results to those entities that satisfy the filter expression.
@@ -251,22 +206,45 @@ func (p *ListEntitiesPager) NextPage(ctx context.Context) (ListEntitiesPageRespo
 // Top: The maximum number of entities that will be returned per page of results.
 // Note: This value does not limit the total number of results if NextPage is called on the returned Pager until it returns false.
 //
-// List returns a Pager, which allows iteration through each page of results. Use nil for listOptions if you want to use the default options.
+// ListEntities returns a Pager, which allows iteration through each page of results. Use nil for listOptions if you want to use the default options.
 // For more information about writing query strings, check out:
 //  - API Documentation: https://docs.microsoft.com/en-us/rest/api/storageservices/querying-tables-and-entities
 //  - README samples: https://github.com/Azure/azure-sdk-for-go/blob/main/sdk/data/aztables/README.md#writing-filters
-func (t *Client) List(listOptions *ListEntitiesOptions) ListEntitiesPager {
+func (t *Client) ListEntities(listOptions *ListEntitiesOptions) *runtime.Pager[ListEntitiesResponse] {
 	if listOptions == nil {
 		listOptions = &ListEntitiesOptions{}
 	}
-	return ListEntitiesPager{
-		client:      t.client,
-		tableName:   t.name,
-		listOptions: listOptions,
-		current:     generated.TableClientQueryEntitiesResponse{},
-		nextPK:      listOptions.PartitionKey,
-		nextRK:      listOptions.RowKey,
-	}
+	return runtime.NewPager(runtime.PageProcessor[ListEntitiesResponse]{
+		More: func(page ListEntitiesResponse) bool {
+			if page.NextPartitionKey == nil || len(*page.NextPartitionKey) == 0 || page.NextRowKey == nil || len(*page.NextRowKey) == 0 {
+				return false
+			}
+			return true
+		},
+		Fetcher: func(ctx context.Context, page *ListEntitiesResponse) (ListEntitiesResponse, error) {
+			var partKey *string
+			var rowKey *string
+			if page != nil {
+				if page.NextPartitionKey != nil {
+					partKey = page.NextPartitionKey
+				}
+				if page.NextRowKey != nil {
+					rowKey = page.NextRowKey
+				}
+			} else {
+				partKey = listOptions.NextPartitionKey
+				rowKey = listOptions.NextRowKey
+			}
+			resp, err := t.client.QueryEntities(ctx, generated.Enum1Three0, t.name, &generated.TableClientQueryEntitiesOptions{
+				NextPartitionKey: partKey,
+				NextRowKey:       rowKey,
+			}, listOptions.toQueryOptions())
+			if err != nil {
+				return ListEntitiesResponse{}, err
+			}
+			return newListEntitiesPage(resp)
+		},
+	})
 }
 
 // GetEntityOptions contains optional parameters for Client.GetEntity
@@ -275,7 +253,7 @@ type GetEntityOptions struct {
 }
 
 func (g *GetEntityOptions) toGenerated() (*generated.TableClientQueryEntityWithPartitionAndRowKeyOptions, *generated.QueryOptions) {
-	return &generated.TableClientQueryEntityWithPartitionAndRowKeyOptions{}, &generated.QueryOptions{Format: generated.ODataMetadataFormatApplicationJSONODataMinimalmetadata.ToPtr()}
+	return &generated.TableClientQueryEntityWithPartitionAndRowKeyOptions{}, &generated.QueryOptions{Format: to.Ptr(generated.ODataMetadataFormatApplicationJSONODataMinimalmetadata)}
 }
 
 // GetEntityResponse contains response fields for Client.GetEntity
@@ -322,8 +300,7 @@ func (t *Client) GetEntity(ctx context.Context, partitionKey string, rowKey stri
 
 // AddEntityOptions contains optional parameters for Client.AddEntity
 type AddEntityOptions struct {
-	// Specifies whether the response should include the inserted entity in the payload. Possible values are return-no-content and return-content.
-	ResponsePreference *ResponseFormat
+	// placeholder for future optional parameters
 }
 
 // AddEntityResponse contains response fields for Client.AddEntityResponse
@@ -355,7 +332,7 @@ func (t *Client) AddEntity(ctx context.Context, entity []byte, options *AddEntit
 	if err != nil {
 		return AddEntityResponse{}, err
 	}
-	resp, err := t.client.InsertEntity(ctx, generated.Enum1Three0, t.name, &generated.TableClientInsertEntityOptions{TableEntityProperties: mapEntity, ResponsePreference: generated.ResponseFormatReturnNoContent.ToPtr()}, nil)
+	resp, err := t.client.InsertEntity(ctx, generated.Enum1Three0, t.name, &generated.TableClientInsertEntityOptions{TableEntityProperties: mapEntity, ResponsePreference: to.Ptr(generated.ResponseFormatReturnNoContent)}, nil)
 	if err != nil {
 		err = checkEntityForPkRk(&mapEntity, err)
 		return AddEntityResponse{}, err
@@ -398,7 +375,7 @@ func (t *Client) DeleteEntity(ctx context.Context, partitionKey string, rowKey s
 // UpdateEntityOptions contains optional parameters for Client.UpdateEntity
 type UpdateEntityOptions struct {
 	IfMatch    *azcore.ETag
-	UpdateMode EntityUpdateMode
+	UpdateMode UpdateMode
 }
 
 func (u *UpdateEntityOptions) toGeneratedMergeEntity(m map[string]interface{}) *generated.TableClientMergeEntityOptions {
@@ -406,7 +383,7 @@ func (u *UpdateEntityOptions) toGeneratedMergeEntity(m map[string]interface{}) *
 		return &generated.TableClientMergeEntityOptions{}
 	}
 	return &generated.TableClientMergeEntityOptions{
-		IfMatch:               to.StringPtr(string(*u.IfMatch)),
+		IfMatch:               to.Ptr(string(*u.IfMatch)),
 		TableEntityProperties: m,
 	}
 }
@@ -416,7 +393,7 @@ func (u *UpdateEntityOptions) toGeneratedUpdateEntity(m map[string]interface{}) 
 		return &generated.TableClientUpdateEntityOptions{}
 	}
 	return &generated.TableClientUpdateEntityOptions{
-		IfMatch:               to.StringPtr(string(*u.IfMatch)),
+		IfMatch:               to.Ptr(string(*u.IfMatch)),
 		TableEntityProperties: m,
 	}
 }
@@ -463,7 +440,7 @@ func updateEntityResponseFromUpdateGenerated(g *generated.TableClientUpdateEntit
 func (t *Client) UpdateEntity(ctx context.Context, entity []byte, options *UpdateEntityOptions) (UpdateEntityResponse, error) {
 	if options == nil {
 		options = &UpdateEntityOptions{
-			UpdateMode: EntityUpdateModeMerge,
+			UpdateMode: UpdateModeMerge,
 		}
 	}
 
@@ -485,7 +462,7 @@ func (t *Client) UpdateEntity(ctx context.Context, entity []byte, options *Updat
 	rowkey := rk.(string)
 
 	switch options.UpdateMode {
-	case EntityUpdateModeMerge:
+	case UpdateModeMerge:
 		resp, err := t.client.MergeEntity(
 			ctx,
 			generated.Enum1Three0,
@@ -496,7 +473,7 @@ func (t *Client) UpdateEntity(ctx context.Context, entity []byte, options *Updat
 			&generated.QueryOptions{},
 		)
 		return updateEntityResponseFromMergeGenerated(&resp), err
-	case EntityUpdateModeReplace:
+	case UpdateModeReplace:
 		resp, err := t.client.UpdateEntity(
 			ctx,
 			generated.Enum1Three0,
@@ -519,9 +496,9 @@ type InsertEntityOptions struct {
 	// ETag is the optional etag for the Table
 	ETag azcore.ETag
 
-	// UpdateMode is the desired mode for the Update. Use ReplaceEntity to replace fields on
-	// the entity, use MergeEntity to merge fields of the entity.
-	UpdateMode EntityUpdateMode
+	// UpdateMode is the desired mode for the Update. Use UpdateModeReplace to replace fields on
+	// the entity, use UpdateModeMerge to merge fields of the entity.
+	UpdateMode UpdateMode
 }
 
 // InsertEntityResponse contains response fields for Client.InsertEntity
@@ -557,16 +534,16 @@ func insertEntityFromGeneratedUpdate(g *generated.TableClientUpdateEntityRespons
 	}
 }
 
-// InsertEntity inserts an entity if it does not already exist in the table. If the entity does exist, the entity is
+// UpsertEntity inserts an entity if it does not already exist in the table. If the entity does exist, the entity is
 // replaced or merged as specified the updateMode parameter. If the entity exists and updateMode is Merge, the property
 // values present in the specified entity will be merged with the existing entity rather than replaced.
 // The response type will be TableEntityMergeResponse if updateMode is Merge and TableEntityUpdateResponse if updateMode is Replace.
 // If the service returns a non-successful HTTP status code, the function returns an *azcore.ResponseError type.
 // Specify nil for options if you want to use the default options.
-func (t *Client) InsertEntity(ctx context.Context, entity []byte, options *InsertEntityOptions) (InsertEntityResponse, error) {
+func (t *Client) UpsertEntity(ctx context.Context, entity []byte, options *InsertEntityOptions) (InsertEntityResponse, error) {
 	if options == nil {
 		options = &InsertEntityOptions{
-			UpdateMode: EntityUpdateModeMerge,
+			UpdateMode: UpdateModeMerge,
 		}
 	}
 	var mapEntity map[string]interface{}
@@ -582,7 +559,7 @@ func (t *Client) InsertEntity(ctx context.Context, entity []byte, options *Inser
 	rowkey := rk.(string)
 
 	switch options.UpdateMode {
-	case EntityUpdateModeMerge:
+	case UpdateModeMerge:
 		resp, err := t.client.MergeEntity(
 			ctx,
 			generated.Enum1Three0,
@@ -593,7 +570,7 @@ func (t *Client) InsertEntity(ctx context.Context, entity []byte, options *Inser
 			&generated.QueryOptions{},
 		)
 		return insertEntityFromGeneratedMerge(&resp), err
-	case EntityUpdateModeReplace:
+	case UpdateModeReplace:
 		resp, err := t.client.UpdateEntity(
 			ctx,
 			generated.Enum1Three0,
@@ -682,9 +659,9 @@ func (t *Client) SetAccessPolicy(ctx context.Context, options *SetAccessPolicyOp
 	return setAccessPolicyResponseFromGenerated(&response), err
 }
 
-// GetTableSASToken is a convenience method for generating a SAS token for a specific table.
+// GetTableSASURL is a convenience method for generating a SAS token for a specific table.
 // It can only be used by clients created by NewClientWithSharedKey().
-func (t Client) GetTableSASToken(permissions SASPermissions, start time.Time, expiry time.Time) (string, error) {
+func (t Client) GetTableSASURL(permissions SASPermissions, start time.Time, expiry time.Time) (string, error) {
 	if t.cred == nil {
 		return "", errors.New("SAS can only be signed with a SharedKeyCredential")
 	}
@@ -697,7 +674,7 @@ func (t Client) GetTableSASToken(permissions SASPermissions, start time.Time, ex
 		StartRowKey:       permissions.StartRowKey,
 		EndPartitionKey:   permissions.EndPartitionKey,
 		EndRowKey:         permissions.EndRowKey,
-	}.NewSASQueryParameters(t.cred)
+	}.Sign(t.cred)
 	if err != nil {
 		return "", err
 	}
@@ -706,6 +683,6 @@ func (t Client) GetTableSASToken(permissions SASPermissions, start time.Time, ex
 	if !strings.Contains(serviceURL, "/") {
 		serviceURL += "/"
 	}
-	serviceURL += t.name + "?" + qps.Encode()
+	serviceURL += t.name + "?" + qps
 	return serviceURL, nil
 }